Remove unnecessary boxing
This shortened some lines so the single-use
variables could be inlined.

+++ java/org/contikios/cooja/Simulation.java
@@ -1040,7 +1040,7 @@ public class Simulation extends Observable implements Runnable {
         speedLimitNone = false;
         speedLimitLastRealtime = System.currentTimeMillis();
         speedLimitLastSimtime = getSimulationTime();
-        speedLimit = newSpeedLimit.doubleValue();
+        speedLimit = newSpeedLimit;
 
         if (delayEvent.isScheduled()) {
           delayEvent.remove();
